Making BFG's VR menus translatable
This PR adds an (incomplete) modification to the menus so the VR-related strings are outputted to the string files.
 - All the files I could think of have been checked and exported.
 - Added a first pass at a Spanish translation of the menus.
 - In some cases, I've duplicated the "Disabled" string into new options. While the official game has one single "Disabled"/"Enabled" string for the entire game, this is a mistake for non-English languages that require adding the gender of the option on each string.

What is currently missing:
 - I am not a decent programmer, so any strings that contained variables that would require extra code in the menus were left intact, though I've already allocated string names for them. You can look for those strings by searching the text ```//VR MOD, STR NAME ASSIGNED```.

Also, would it be possible to add hints, like the ones at the main Options screen, at all the VR options? I think I (or someone else) could rework a bit the information that's located at the readme file so people would know what does each option have without having to leave the game.